Initial implementation for a roll dice endpoint
This commit adds an action that processes and rolls dices.  It can be
given a query command such as "1d10" or "2d5,3d10", and it will roll
dices to honor the request.

The command is provided as the ?q query parameter for the /roll
endpoint.  There are validators that test that the given query conforms
to the rules (i.e. it doesn't have weird characters and it's actually
present).

Some important considerations taken from #1:

* Dices have to be rolled one by one.  So a command such as 2d10 must
  actually generate two different outcomes in range [1,10].  You must not
  generate a single outcome in range [2,20], that would be an error.
* You can use a weird and phisically impossible number of faces for a
  dice, not just the common 4, 6, 8, 10, 12 or 20 dices.  You can roll a
  23 faces dice if you wish.

Some additional considerations:

* I don't currently care whether you write d or D.