feat(spells,frontend): add simple spell searcher
This searcher employs autocompletion with NO request debouncing
to query the server for spells matching the terms written. This
is done becuase the way FTS respond with data is not intuitive
and thus, forcing the user the wait the cooldown interval to check
if their terms yielded results, would degrade UX.

This version of the component is not even reusable and quite ugly
but it gets the job done.