feat: add endpoints for JWT authentication
This commit adds a JWT based authentication middleware through the
jwt-auth library, as well as controllers to issue, refresh and expire
JWT tokens.

The goal is to have an authentication system that is true stateless
that can be used in frontend, even when the frontend is served from a
different domain where the auth:api middleware can't be used due to
cross-origin cookie issues.

This way, frontend can now have a login view that will ask the backend
to issue a JWT token. Frontend will need to user the JWT token using
Bearer Authentication in following requests, in order for the backend
to identify the user and be truthful alternative to the cookie
authentication, now in use.

To issue JWT tokens, it will be necessary to have a private key that
will allow to sign these tokens. To generate the proper environment
variable, when the repository is set up for development or
deployment, one can use the `php artisan jwt:secret` command to
generate a secret, that will be added to the .env file.

Exposed API methods:

* POST /api/auth/sessions – issue a JWT token.
* GET /api/auth/session – validate a JWT token and provide user info.
* DELETE /api/auth/session — expire the previously issued JWT token.
* POST /api/auth/refresshments – refresh a JWT token and issue another
  one.

Issues to solve after this PR:

* Once OpenRPG-UI has a login system that uses these endpoints, all
the frontend functionality done in OpenRPG will be replicated in
OpenRPG-UI, so we can already consider deprecating the current Vue
frontend in OpenRPG and treat OpenRPG as a REST API only repository.

* Once the Vue frontend provided by OpenRPG is deprecated, we can
evaluate removing the auth:api middleware and renaming auth:jwt to
auth:api.

Closes #49