fix: model tests should use database transactions
This commit will make the model tests use the DatabaseTransactions
trait, which forces the tests to actually cleanup after running.
Tests that are currently using RefreshDatabase should refrain from
using it and switch to DatabaseTransactions instead, to avoid removing
seeds that may be expected to be present in development or testing.